convert playlist slice to queries/mutations

+++ src/Data/Models/Playlist.cs
@@ -2,19 +2,22 @@ using System.Text;
 
 namespace PlaylistManager.Data.Models;
 
-public class Playlist {
+public class Playlist
+{
     public int id { get; set; }
     public required string title { get; set; }
     public List<Video> entries { get; set; } = new List<Video>();
     public DateTime createdAt { get; set; }
     public const string M3U_NEWLINE = "\r\n";
 
-    public string ToM3U() {
+    public string ToM3U()
+    {
         var sb = new StringBuilder();
         sb.Append("#EXTM3U");
         sb.Append(M3U_NEWLINE);
 
-        foreach(var video in entries) {
+        foreach (var video in entries)
+        {
             sb.Append($"#EXTINF:{video.duration},{video.title}{M3U_NEWLINE}{video.filename}{M3U_NEWLINE}");
         }
 
@@ -23,6 +26,6 @@ public class Playlist {
 }
 
 public record PlaylistCreateRequest(string title);
-public record PlaylistEntryChangeRequest(int videoId, int entryorder);
+public record PlaylistEntryChangeRequest(long videoId, int entryorder);
 
-public record PlaylistEntriesUpdateRequest(List<PlaylistEntryChangeRequest> toAdd, List<int> toRemove, List<PlaylistEntryChangeRequest> toUpdate);
+public record PlaylistEntriesUpdateRequest(List<PlaylistEntryChangeRequest> toAdd, List<long> toRemove, List<PlaylistEntryChangeRequest> toUpdate);
